+++ CHANGELOG.md
@@ -5,10 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.1]
 
 - Added first two functions (only docstrings): get_usr_input.py and sentiment_analysis.py
 - Added fifth function (only docstring): translate_sentence.py
 - Added third function (only docstring): fetch_definition.py
 - Added sixth function (only docstring): wordcloud.py
 - Added fourth function (only docstring): word_replacement.py
+
+
+## [1.0.0]
+
+- Added definition fetch code and unit tests: test_fetch_definition.py and fetch_definition.py
+- Added sentiment analysis code and unit tests: test_sentiment_analysis.py and sentiment_analysis.py
+- Added function logic to replace a word in a sentence and compute the sentiment in word_replacement.py
+- Added helper function sentiment_specifics in utils.py to provide detailed sentiment analysis results.